fix(web): render /organization settings without slug redirect
Eliminate the Settings double hop by sharing org settings content between the context path and deep-link slug route.
