+++ apps/web/src/app/(app)/organization/page.test.ts
@@ -5,14 +5,15 @@ import { describe, expect, it } from "vitest";
 
 const dir = dirname(fileURLToPath(import.meta.url));
 
-describe("/organization route alias", () => {
-  it("redirects to active org slug detail or home", () => {
+describe("/organization route", () => {
+  it("renders active-org settings in place without hopping to the slug route", () => {
     const source = readFileSync(join(dir, "page.tsx"), "utf8");
 
-    expect(source).toContain("userService.getActiveOrganization()");
-    expect(source).toContain(
-      "`/organizations/${encodeURIComponent(organization.slug)}`",
-    );
+    expect(source).toContain("OrganizationSettingsContent");
+    expect(source).toContain("userService.getActiveOrganizationId()");
+    expect(source).toContain("userService.getMyMembersWithOrganizations()");
     expect(source).toContain('redirect("/")');
+    expect(source).not.toContain("/organizations/${");
+    expect(source).not.toContain("getActiveOrganization()");
   });
 });
